- sync to develop - fix bug that caused by parser cannot find stuffs corrupting other part of the code - fix crash because the stack return is not properly set

+++ src/Ext/CaptureManager/Body.cpp
@@ -239,7 +239,7 @@ void FakeCaptureManagerClass::__DecideUnitFate(TechnoClass* pTechno , bool Captu
     
     if (availableMoney >= RulesClass::Instance->AICaptureLowMoneyMark) {
         if (this->Owner->Owner->GetPowerPercentage() >= 1.0) {
-            float healthRatio = pTechno->GetHealthRatio();
+            float healthRatio = (float)pTechno->GetHealthRatio();
             
             if (healthRatio >= RulesClass::Instance->AICaptureWoundedMark) {
                 nCondition = 3;
